requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
stripe==5.0.0
python-dotenv==0.19.0
jinja2==3.1.2
requests==2.28.2
pydantic==1.10.4
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
python-decouple==3.7
SQLAlchemy-Utils==0.39.0